Remove old auth examples
We shall maintain only Python examples to the auth procedures.
Relevant docs are updated accordingly.
