+++ .gitignore
@@ -1,7 +1,32 @@
-/venv/
+# Python
+__pycache__/
+*.py[cod]
+*$py.class
+*.so
+
+# Virtual environments
+.venv/
+venv/
+ENV/
+
+# IDE
 .idea/
+.vscode/
+*.swp
+*.swo
+
+# OS
 .DS_Store
+Thumbs.db
+
+# Local config files with credentials
+.env
+*.local.py
+
+# Token files
 access_token.json
 .access_token.json
+
+# Node.js
 node_modules/
 package-lock.json
