+++ auth/py-oauth1/requirements.txt
@@ -1,3 +1,2 @@
 requests
 requests-oauthlib
-oauth2
