--- examples/readme.md
@@ -1,50 +1,0 @@
-**Disclaimer**: Planview provides these examples for instructional purposes. While you are welcome to use this
-code in any way you see fit - Planview does not accept any liability or responsibility for you choosing to do so.
-
-# API Code Examples
-
-This directory contains examples of API usage across the Planview ProjectPlace product.
-
-## Getting Started: Authentication
-
-**NEW!** Before using any of these examples, start with our authentication guides:
-
-👉 **[Authentication Overview (AUTH_README.md)](./AUTH_README.md)** - Choose the right auth method for your needs
-
-### Quick Links to Authentication Examples:
-- **[OAuth2 Authorization Code Flow](./py-oauth2-authorization-code/)** ⭐ Recommended for user access
-- **[OAuth2 Client Credentials Flow](./py-oauth2-client-credentials/)** ⭐ Recommended for service accounts  
-- **[OAuth1 (Legacy)](./py-oauth1/)** - For maintaining existing integrations
-
-## Available Examples
-
-### Authentication Examples
-- **py-oauth2-authorization-code** - OAuth2 user authentication flow
-- **py-oauth2-client-credentials** - OAuth2 robot/service account authentication  
-- **py-oauth1** - Legacy OAuth1 authentication
-
-### Data Operations Examples
-- **py-board-webhooks** - Set up and manage board webhooks
-- **py-bulk-update-emails** - Bulk update user email addresses
-- **py-consume-odata** - Access and download OData feeds
-- **py-download-archived-workspaces** - Download archived workspace data
-- **py-download-document** - Download documents from workspaces
-- **py-enforce-column-name** - Bulk rename board columns across workspaces
-- **py-list-document-archive** - List document archive contents
-- **py-remove-inactive-users** - Remove inactive users from account
-- **py-upload-document** - Upload documents to workspaces
-- **node-js-import-cards-with-excel** - Import cards from Excel spreadsheets
-
-## About These Examples
-
-* Examples are written in a specific language - designated by the prefixes e.g `py-` for Python and `node-js` for Node etc.
-* The code herein is meant to be possible to run successfully with minor modifications for authentication.
-* All examples include README files with setup instructions and usage examples.
-* As the disclaimer above states: while we encourage you to study the code to understand our APIs - we do not accept responsibility for you running or modifying the code.
-
-## Resources
-
-- [API Documentation](https://api.projectplace.com/apidocs)
-- [OAuth2 Guide](https://api.projectplace.com/apidocs#articles/pageOAuth2.html)
-- [Success Center](https://success.planview.com/Planview_ProjectPlace)
-
