--- oauth2/.npmrc
@@ -1,2 +1,0 @@
-# .npmrc
-engine-strict=true
